Add files via upload
Q1
? Problem Statement:
Write a function to reverse a singly linked list. The function should take the head of the linked list as an input and return the head of the reversed list.

1> Approach:
To reverse a linked list, you need to reverse the pointers of each node. You can use three pointers: prev, current, and next.

1.} Initialize: Set prev to null and current to head.
2.} Iterate: Traverse the list, changing the next pointer of each node to point to the previous node.
3.} Update Pointers: Move prev to current and current to next until you reach the end of the list.
4.} Set New Head: The new head of the reversed list will be the last non-null prev,.


Q2
? Problem Statement:
Given a singly linked list, find the K-th to last element. If K is 1, the result is the last element; if K is 2, the result is the second-to-last element, and so on.

<> Approach:
To solve this problem, we can use the two-pointer technique, often referred to as the fast and slow pointer approach:

1.} Initialize Two Pointers: Start both pointers, fast and slow, at the head of the list.
2} Move Fast Pointer: Move the fast pointer K steps ahead.
3} Move Both Pointers: Move fast and slow pointers one step at a time until fast reaches the end of the list.
4.}Result: The slow pointer will now point to the K-th to the last element.
This approach ensures a single pass through the list, achieving O(n) time complexity.